test: getter 사용을 제거하면서 LottoBalls의 책임이 된 기능 테스트 이동.
- 테스트 실행중 shuffle이 먼저 실행되면 로또 번호 범위 검증 테스트가 깨져 수정함.
